Added the first pop up window
I added the pop up window that appears on the screen when the user wants to add a new food category. I also added some comments on the java files and started to code how the data will be modeled (only an enum class is present at the moment). We need this in order to add the pop up window's functionality.