Pop up window works
The pop up window is functional - i.e. you can fill in the fields, cancel the procedure, tap on add and the window will disapper if all fields are filled, otherwise it will notify the user to do so. Now a message appears on the inventory if there are no lists. I also started working with how data will be stored and RecyclerView, but they are not finished yet.